+++ chess_game/Piece.h
@@ -11,16 +11,16 @@ class Board;
 class Piece {
   public:
     // Knight move
-    static constexpr int knight_dx[8] = {2, 2, -2, -2, 1, 1, -1, -1};
-    static constexpr int knight_dy[8] = {1, -1, 1, -1, 2, -2, 2, -2};
+    static constexpr int8_t knight_dx[8] = {2, 2, -2, -2, 1, 1, -1, -1};
+    static constexpr int8_t knight_dy[8] = {1, -1, 1, -1, 2, -2, 2, -2};
 
     // Piece type:
     PieceType type;
     // Color of the piece:
     bool color; // 0 for white, 1 for black
     // X and Y coordinates of the piece: (-1, -1) for piece off grid
-    int x;
-    int y;
+    int8_t x;
+    int8_t y;
     // Pawn: Double move flag - 1 if this pawn can move two squares
     bool double_move;
 
@@ -28,19 +28,19 @@ class Piece {
 
     bool get_color();
 
-    int get_x();
+    int8_t get_x();
 
-    int get_y();
+    int8_t get_y();
 
     bool get_double_move();
 
     // Function that returns x,y coordinates of all possible moves
 
-    std::vector<std::pair<std::pair<int, int>, std::pair<int, int>>> get_possible_moves(Board* board) const;
+    std::vector<std::pair<std::pair<int8_t, int8_t>, std::pair<int8_t, int8_t>>> get_possible_moves(Board* board) const;
 
     // Constructor
 
-    Piece(PieceType new_type, bool new_color, int new_x, int new_y);
+    Piece(PieceType new_type, bool new_color, int8_t new_x, int8_t new_y);
 
 };
